refactor: move HeaderBar and Profile up one level
